+++ lib/common/singletons/current_user.dart
@@ -21,7 +21,7 @@ import 'package:flutter/foundation.dart';
 import '../../get_it.dart';
 import '../../manager/address_manager.dart';
 import '../../manager/favorites_manager.dart';
-import '../../repository/user_repository.dart';
+import '../../repository/parse_server/user_repository.dart';
 import '../models/address.dart';
 import '../models/user.dart';
