+++ lib/features/signup/signup_controller.dart
@@ -21,7 +21,7 @@ import '../../common/models/user.dart';
 import '../../common/singletons/app_settings.dart';
 import '../../components/custon_field_controllers/masked_text_controller.dart';
 import '../../get_it.dart';
-import '../../repository/user_repository.dart';
+import '../../repository/parse_server/user_repository.dart';
 import 'signup_state.dart';
 
 class SignupController extends ChangeNotifier {
