+++ lib/manager/favorites_manager.dart
@@ -23,7 +23,7 @@ import '../common/models/ad.dart';
 import '../common/models/favorite.dart';
 import '../common/singletons/current_user.dart';
 import '../get_it.dart';
-import '../repository/favorite_repository.dart';
+import '../repository/parse_server/favorite_repository.dart';
 
 class FavoritesManager {
   final List<FavoriteModel> _favs = [];
