+++ lib/manager/state_manager.dart
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with bgbazzar.  If not, see <https://www.gnu.org/licenses/>.
 
-import '../repository/ibge_repository.dart';
+import '../repository/gov_api/ibge_repository.dart';
 import '../common/models/state.dart';
 
 /// This class provides a singleton manager for handling states in Brazil.
