+++ lib/repository/parse_server/address_repository.dart
@@ -19,7 +19,7 @@ import 'dart:developer';
 
 import 'package:parse_server_sdk_flutter/parse_server_sdk_flutter.dart';
 
-import '../common/models/address.dart';
+import '../../common/models/address.dart';
 import 'common/constants.dart';
 import 'common/parse_to_model.dart';
