+++ lib/repository/parse_server/boardgame_repository.dart
@@ -20,7 +20,7 @@ import 'dart:io';
 
 import 'package:parse_server_sdk_flutter/parse_server_sdk_flutter.dart';
 
-import '../common/models/boardgame.dart';
+import '../../common/models/boardgame.dart';
 import 'common/constants.dart';
 import 'common/parse_to_model.dart';
