+++ lib/repository/parse_server/favorite_repository.dart
@@ -19,8 +19,8 @@ import 'dart:developer';
 
 import 'package:parse_server_sdk_flutter/parse_server_sdk_flutter.dart';
 
-import '../common/models/ad.dart';
-import '../common/models/favorite.dart';
+import '../../common/models/ad.dart';
+import '../../common/models/favorite.dart';
 import 'common/constants.dart';
 import 'common/parse_to_model.dart';
