+++ lib/repository/sqlite/mechanic_repository.dart
@@ -17,8 +17,8 @@
 
 import 'dart:developer';
 
-import '../common/models/mechanic.dart';
-import '../store/mech_store.dart';
+import '../../common/models/mechanic.dart';
+import 'store/mech_store.dart';
 
 /// This class provides methods to interact with the Parse Server
 /// to retrieve a list of mechanics.
