+++ lib/repository/sqlite/store/bgg_rank_store.dart
@@ -17,26 +17,23 @@
 
 import 'dart:developer';
 
-import '../get_it.dart';
+import '../../../get_it.dart';
 import 'constants/constants.dart';
 import 'database_manager.dart';
 
-class MechStore {
+class BggRankStore {
   static final _databaseManager = getIt<DatabaseManager>();
 
-  static Future<List<Map<String, dynamic>>> queryMechs(
-      [String langCode = 'pt_BR']) async {
+  static Future<List<Map<String, dynamic>>> queryRankGameNames(int year) async {
     final database = await _databaseManager.database;
 
-    final getColumns = (langCode != 'pt_BR')
-        ? [mechId, mechName, mechDescription]
-        : [mechId, mechNome, mechDescricao];
-    final orderByName = (langCode != 'pt_BR') ? mechName : mechNome;
     try {
       List<Map<String, dynamic>> result = await database.query(
-        mechTable,
-        columns: getColumns,
-        orderBy: orderByName,
+        rankTable,
+        columns: [rankId, rankGameName],
+        where: '$rankYearPublished > ?',
+        whereArgs: [year],
+        orderBy: rankGameName,
       );
 
       return result;
@@ -46,23 +43,18 @@ class MechStore {
     }
   }
 
-  static Future<String> queryDescription(int id,
-      [String langCode = 'pt_BR']) async {
+  static Future<Map<String, dynamic>> queryRankFromId(int id) async {
     final database = await _databaseManager.database;
-    final getColumn = (langCode != 'pt_BR') ? mechDescription : mechDescricao;
-    try {
-      final result = await database.query(
-        mechTable,
-        columns: [getColumn],
-        where: '$mechId = ?',
-        whereArgs: [id],
-      );
 
-      if (result.isEmpty) return '';
-      return result[0][getColumn] as String;
+    try {
+      List<Map<String, dynamic>> result = await database
+          .query(rankTable, where: '$rankId = ?', whereArgs: [id]);
+
+      if (result.isEmpty) return {};
+      return result.first;
     } catch (err) {
       log('Error: $err');
-      return '';
+      return {};
     }
   }
 }
