+++ lib/repository/sqlite/store/constants/constants.dart
