+++ lib/repository/sqlite/store/mech_store.dart
@@ -17,23 +17,26 @@
 
 import 'dart:developer';
 
-import '../get_it.dart';
+import '../../../get_it.dart';
 import 'constants/constants.dart';
 import 'database_manager.dart';
 
-class BggRankStore {
+class MechStore {
   static final _databaseManager = getIt<DatabaseManager>();
 
-  static Future<List<Map<String, dynamic>>> queryRankGameNames(int year) async {
+  static Future<List<Map<String, dynamic>>> queryMechs(
+      [String langCode = 'pt_BR']) async {
     final database = await _databaseManager.database;
 
+    final getColumns = (langCode != 'pt_BR')
+        ? [mechId, mechName, mechDescription]
+        : [mechId, mechNome, mechDescricao];
+    final orderByName = (langCode != 'pt_BR') ? mechName : mechNome;
     try {
       List<Map<String, dynamic>> result = await database.query(
-        rankTable,
-        columns: [rankId, rankGameName],
-        where: '$rankYearPublished > ?',
-        whereArgs: [year],
-        orderBy: rankGameName,
+        mechTable,
+        columns: getColumns,
+        orderBy: orderByName,
       );
 
       return result;
@@ -43,18 +46,23 @@ class BggRankStore {
     }
   }
 
-  static Future<Map<String, dynamic>> queryRankFromId(int id) async {
+  static Future<String> queryDescription(int id,
+      [String langCode = 'pt_BR']) async {
     final database = await _databaseManager.database;
-
+    final getColumn = (langCode != 'pt_BR') ? mechDescription : mechDescricao;
     try {
-      List<Map<String, dynamic>> result = await database
-          .query(rankTable, where: '$rankId = ?', whereArgs: [id]);
+      final result = await database.query(
+        mechTable,
+        columns: [getColumn],
+        where: '$mechId = ?',
+        whereArgs: [id],
+      );
 
-      if (result.isEmpty) return {};
-      return result.first;
+      if (result.isEmpty) return '';
+      return result[0][getColumn] as String;
     } catch (err) {
       log('Error: $err');
-      return {};
+      return '';
     }
   }
 }
