+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: bgbazzar
 description: "A new Flutter project."
 publish_to: 'none'
-version: 0.6.11+34
+version: 0.6.12+35
 
 environment:
   sdk: '>=3.4.1 <4.0.0'
